Rename namespace and license
