+++ source/RequestResponse/RequestResponse.csproj
@@ -4,10 +4,10 @@
         <TargetFramework>net6.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <AssemblyName>Energinet.DataHub.MeteringPoints.ResponseRequest</AssemblyName>
-        <RootNamespace>Energinet.DataHub.MeteringPoints.ResponseRequest</RootNamespace>
-        <PackageId>Energinet.DataHub.MeteringPoints.ResponseRequest</PackageId>
-        <Title>EnergySupplying IntegrationEvents library</Title>
+        <AssemblyName>Energinet.DataHub.MeteringPoints.RequestResponse</AssemblyName>
+        <RootNamespace>Energinet.DataHub.MeteringPoints.RequestResponse</RootNamespace>
+        <PackageId>Energinet.DataHub.MeteringPoints.RequestResponse</PackageId>
+        <Title>MeteringPoint RequestResponse library</Title>
         <Authors>Energinet-DataHub</Authors>
         <Description>Library to use for implementation of request response integration from EnergySupplying (MarketRoles) domain.</Description>
         <PackageProjectUrl>https://github.com/Energinet-DataHub</PackageProjectUrl>
@@ -18,7 +18,7 @@
       [Documentation](https://github.com/Energinet-DataHub/geh-metering-point/blob/master/source/RequestResponse/documents/development.md)
     </PackageReleaseNotes>
         <Company>Energinet-DataHub</Company>
-        <Product>Energinet.DataHub.MeteringPoints.ResponseRequest</Product>
+        <Product>Energinet.DataHub.MeteringPoints.RequestResponse</Product>
     </PropertyGroup>
 
 </Project>
