--- source/ResponseRequest/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace ResponseRequest;
-
-public class Class1
-{
-}
